Move toHex to a seperate file
